\name{enzact}
\alias{enzact}
\docType{data}
\title{
Biomarkers activity values with different treatments
}
\description{
This data set gives level or activity of 28 biomarkers, generated as simulated values.
}
\usage{data("enzact")}
\format{
  A data frame with 28 observations on the following 8 variables.
  \describe{
    \item{\code{treatment}}{a character vector}
    \item{\code{biomarker1}}{a numeric vector}
    \item{\code{biomarker2}}{a numeric vector}
    \item{\code{biomarker3}}{a numeric vector}
    \item{\code{biomarker4}}{a numeric vector}
    \item{\code{biomarker5}}{a numeric vector}
  }
}
\examples{
data(enzact)

}
\keyword{datasets}
