\name{enzact2}
\alias{enzact2}
\docType{data}
\title{
Biomarkers activity values with different treatments
}
\description{
This data set gives level or activity of 38 biomarkers, generated as simulated values.
}
\usage{data("enzact2")}
\format{
  A data frame with 38 observations on the following 6 variables.
  \describe{
    \item{\code{sites}}{a character vector}
    \item{\code{biomarker1}}{a numeric vector}
    \item{\code{biomarker2}}{a numeric vector}
    \item{\code{biomarker3}}{a numeric vector}
    \item{\code{biomarker4}}{a numeric vector}
    \item{\code{biomarker5}}{a numeric vector}
  }
}
\examples{
data(enzact2)
## maybe str(enzact2) ; plot(enzact2) ...
}
\keyword{datasets}
