% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibrv2_std.R
\name{ibrv2_std}
\alias{ibrv2_std}
\title{Standardized values for IBRv2 (Integrated Biological Responses version 2)}
\usage{
ibrv2_std(df, na.rm = TRUE, ...)
}
\arguments{
\item{df}{A data frame containing values the enzymes activities with a reference value on the first rows.}
}
\description{
This function calculates de standardized values for IBRv2 index purposed by Sanchez et al. 2013.
}
\details{
Gives a dataframe with the standardized values for each area/treatment in your input data to be used to create the radarchart with the function ibrv2_chart from this same package.
}
\section{References}{


Sanchez, W., Burgeot, T., & Porcher, J.-M. (2013). A novel “Integrated Biomarker Response” calculation based on reference deviation concept. Environmental Science and Pollution Research, 20(5), 2721–2725. https://doi.org/10.1007/s11356-012-1359-1
}

\examples{
data(enzact2)

ibrv2_std(enzact2)

}
