% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBRtools-package.R
\docType{package}
\name{IBRtools}
\alias{IBRtools}
\title{IBRtools: Indexes for biomonitoring assessment and graphic creation}
\description{
\if{html}{\figure{r-ibr-logo.png}{options: align='right' alt='logo' width='120'}}
}
\details{
The IBRtools package calculates two important indexes (IBR \link{Integrated Biomarker Response} and IBRv2 \link{Integrated Biological Response version 2}), it also calculates the standardized values for enzyme activity for each index, and it has a graphing function to perform radarplots that make great data visualization for this type of data!

This packege also comes with 3 example datasets: enzact, enzact_coef and enzact2

You can use them to practice or to see how the functions work so you can apply in your own datasets!

To load them, just use the command: data(nameofdataset)
}
\section{IBRtools functions}{

There are 6 functions in this package, three for each index.

IBR: ibr_index, ibr_std, ibr_chart

IBRv2: ibrv2_index, ibrv2_bdi, ibrv2_chart
}

\section{Authors}{

Author, Mainteiner: Anna Carolina Resende

\code{\email{annac.resende@gmail.com}}

\code{\url{https://orcid.org/0000-0003-0909-1255}}

Author: Diego Mauro Carneiro Pereira

\code{\email{diegom8135@gmail.com}}

\code{\url{https://orcid.org/0000-0003-1554-3684}}
}

